+++ src/main/java/org/neos/lambda/sort/OrderNamesAlphabetical.java
@@ -1,4 +1,4 @@
-package org.neos.lambda.order;
+package org.neos.lambda.sort;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -10,16 +10,16 @@ import java.util.List;
  * 15/06/2018
  *
  */
-public class OrderNamesLenght {
+public class OrderNamesAlphabetical {
 
 	public static void main(String[] args) {
-		
 		List<String> names = Arrays.asList("hugo","victoria","veronica","mario","janet","alejandra","roberto");
-		Collections.sort(names, (cad1,cad2)-> cad1.length()-cad2.length());
+		Collections.sort(names,(cad1, cad2) -> cad1.compareTo(cad2));
 		
 		for(String name: names) {
 			System.out.println("name: " + name);
 		}
+		
 	}
 
 }
