+++ src/main/java/org/neos/lambda/sort/OrderNamesAlphabeticalMethod.java
@@ -1,8 +1,7 @@
-package org.neos.lambda.order;
+package org.neos.lambda.sort;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 
 /**
@@ -11,16 +10,16 @@ import java.util.List;
  * 15/06/2018
  *
  */
-public class OrderNamesLenghtMethod {
+public class OrderNamesAlphabeticalMethod {
 
 	public static void main(String[] args) {
-		
 		List<String> names = Arrays.asList("hugo","victoria","veronica","mario","janet","alejandra","roberto");
-		Collections.sort(names,Comparator.comparing(String::length));
+		Collections.sort(names, String::compareToIgnoreCase);
 		
 		for(String name: names) {
 			System.out.println("name: " + name);
 		}
+		
 	}
 
 }
