+++ src/main/java/org/neos/lambda/sort/OrderNamesLenghtMethod.java
@@ -1,7 +1,8 @@
-package org.neos.lambda.order;
+package org.neos.lambda.sort;
 
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 /**
@@ -10,16 +11,16 @@ import java.util.List;
  * 15/06/2018
  *
  */
-public class OrderNamesAlphabetical {
+public class OrderNamesLenghtMethod {
 
 	public static void main(String[] args) {
+		
 		List<String> names = Arrays.asList("hugo","victoria","veronica","mario","janet","alejandra","roberto");
-		Collections.sort(names,(cad1, cad2) -> cad1.compareTo(cad2));
+		Collections.sort(names,Comparator.comparing(String::length));
 		
 		for(String name: names) {
 			System.out.println("name: " + name);
 		}
-		
 	}
 
 }
